"Post merge commit"

+++ core/src/com/fixthewall/logic/WallLogic.java
@@ -23,11 +23,11 @@ public class WallLogic {
         return health;
     }
 
-    public void setHealth(int health) {
+    public void setHealth(float health) {
         this.health = health;
     }
 
-    public int getMaxHealth() {
+    public float getMaxHealth() {
         return maxHealth;
     }
